Moved type parsing to a separate module
The parser will be integrated with tyhandler.
